% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-predictions.R
\name{calcPrediction}
\alias{calcPrediction}
\title{Calculate River Invertebrate Classification Tool (RICT) predictions}
\usage{
calcPrediction(observed_values)
}
\arguments{
\item{observed_values}{Dataframe of observed environmental values
\describe{
  \item{SITE}{Site identifier}
  \item{Waterbody}{Water body identifier}
  ...
}}
}
\value{
Dataframe of predicted biotic scores and probility of observed values
falling into each statistical grouping of rivers.
}
\description{
Calculate River Invertebrate Classification Tool (RICT) predictions
}
\examples{
\dontrun{
predictions <- calcPrediction(observed_values = rict::demo_observed_values)
}
}
