% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-predictions.R
\name{calcPredictions}
\alias{calcPredictions}
\title{Calculate River Invertebrate Classification Tool (RICT) predictions}
\usage{
calcPredictions(observed_values)
}
\arguments{
\item{observed_values}{Dataframe of observed environmental values}
}
\value{
Dataframe of predicted biotic scores and probility of observed values
falling into each statistical grouping of rivers.
}
\description{
Calculate River Invertebrate Classification Tool (RICT) predictions
}
\examples{
\dontrun{
predictions <- calcPredictions(observed_values = rict::demo_observed_values)
}
}
