% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRict.R
\name{calcRict}
\alias{calcRict}
\title{Calculate River Invertebrate Classification Tool (RICT) predictions}
\usage{
calcRict(observed_values, expected_biotic_scores)
}
\arguments{
\item{observed_values}{Dataframe of observed environmental values}

\item{expected_biotic_scores}{Dataframe of expected biotic scores for each
statistical grouping of rivers (referred to as 'end groups means').}
}
\value{
Dataframe of predicted biotic scores and probility of observed values
falling into each statistical grouping of rivers.
}
\description{
Calculate River Invertebrate Classification Tool (RICT) predictions
}
\examples{
\dontrun{
observed_values <- read.csv(file="inst/extdat/fba-test-whpt.csv")
expected_biotic_scores <- read.csv(file="inst/extdat/expected_biotic_scores.csv")
output <- calcRict(observed_values = observed_values,
                   expected_biotic_scores = expected_biotic_scores)
}
}
