% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-probability.R
\name{compare_probability}
\alias{compare_probability}
\title{Compare Proportion of Results in each Class}
\usage{
compare_probability(
  a = NULL,
  b = NULL,
  eqr_bands = c(0, 0.2, 0.4, 0.6, 0.8, 1),
  cap_eqrs = T,
  labels = 5:1
)
}
\arguments{
\item{a}{numeric list of results 'a'}

\item{b}{numeric list of results 'b'}

\item{eqr_bands}{Class boundaries default is eqr_bands = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0)}

\item{cap_eqrs}{Optional TRUE/FALSE - by default results are capped between 0-1}

\item{labels}{Optional List of labels for labelling each class}
}
\value{
returns dataframe
}
\description{
Compares proportion of results 'a' falling into the same or different class
as results 'b'. Designed for comparing simulated Environmental Quality Ratios
(EQR). It is a low-level function used by the higher-level
\code{\link{rict_compare}} function to compare the EQRs of two paired
results.
}
\examples{
\dontrun{
test <- compare_probability(
  a = c(0.3, 0.4, 0.6),
  b = c(1, 0.9, 0.6),
  eqr_bands <- c(0, 0.47, 0.56, 0.68, 0.8, 1)
)
}
}
