% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-test.R
\name{compare_test}
\alias{compare_test}
\title{A Statistical Test to Compare Two Results}
\usage{
compare_test(a, b)
}
\arguments{
\item{a}{numeric results 'a'}

\item{b}{numeric reuslts 'b'}
}
\value{
dataframe containing stats
}
\description{
A Statistical Test to Compare Two Results
}
\examples{
\dontrun{
test <- compare_test(
  a = c(0.3, 0.4, 0.6),
  b = c(1, 0.9, 0.6)
)
}
}
