% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_color_party.R
\name{data_color_party}
\alias{data_color_party}
\title{Color Columns with Partisan Scales}
\usage{
data_color_party(tab, columns = gt::everything(), ...)
}
\arguments{
\item{tab}{A \code{gt} table with class \code{gt_tbl}}

\item{columns}{the columns to color with partisan colors}

\item{...}{additional arguments passed on to \code{gt::data_color()}}
}
\value{
A \link[gt:gt]{gt::gt}
}
\description{
Color Columns with Partisan Scales
}
\examples{
rict(wv_plans, 'cd_2020') \%>\%
  data_color_party(columns = 'e_dvs')
}
