% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-gis-values-log.R
\docType{data}
\name{demo_gis_values_log}
\alias{demo_gis_values_log}
\title{Demo observed values for gis model}
\format{
An object of class \code{data.frame} with 24 rows and 36 columns.
}
\usage{
demo_gis_values_log
}
\description{
A dataset containing demo observed log transformed gis values for model 44.
Stored as .csv:
}
\keyword{datasets}
