% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_observed_values.R
\docType{data}
\name{demo_observed_values}
\alias{demo_observed_values}
\title{Demo observed values}
\format{A data frame with 24 rows and 33 variables:
\describe{
  \item{SITE}{Site identifier}
  \item{Waterbody}{Water body identifier}
  \item{Year}{Year of sampling}
  \item{NGR}{National grid letters}
  \item{Easting}{Easting co-ordinate}
  \item{Northing}{Northing co-ordinate}
  \item{Altitude}{Altitude in metres}
  \item{Slope}{Slope in metres per km}
  \item{Discharge}{Discharge categories}
  \item{Velocity}{Velocity in metres per second}
  \item{Dist_from_Source}{Dist_from_Source in km}
  \item{Mean_Width}{Mean Width in metres}
  \item{Mean_Depth}{Mean Width in metres}
  \item{Alkalinity}{Hardness, Calcium concentration or Electrical conductivity can be provided
  if unable to provide Alkalinity – One of the values must be present (spreadsheet validation)}
  \item{Boulder_Cobbles}{Substrate}
  \item{Pebbles_Gravel}{Substrate}
  \item{Sand}{Substrate}
  \item{Silt_Clay}{Substrate}
  \item{Hardness}{Can be provided instead of Alkalinity and follows the same validation rules (spreadsheet validation)}
  \item{Calcium}{Can be provided instead of Alkalinity and follows the same validation rules (spreadsheet validation)}
  \item{Conductivity}{Can be provided instead of Alkalinity and follows the same validation rules (spreadsheet validation)}
  \item{Spr_Season_ID}{season ID}
  \item{Spr_TL2_WHPT_ASPT [AbW,DistFam]}{ASPT}
  \item{Spr_TL2_WHPT_NTaxa \(AbW,DistFam\)}{NTAXA}
  \item{Spr_Ntaxa_Bias}{bias}
  \item{Sum_Season_ID}{season ID}
  \item{Sum_TL2_WHPT_ASPT \(AbW,DistFam\)}{ASPT}
  \item{Sum_TL2_WHPT_NTaxa \(AbW,DistFam\)}{NTAXA}
  \item{Sum_Ntaxa_Bias}{bias}
  \item{Aut_Season_ID}{season ID}
  \item{Aut_TL2_WHPT_ASPT \(AbW,DistFam\)}{ASPT}
  \item{Aut_TL2_WHPT_NTaxa \(AbW,DistFam\)}{NTAXA}
  \item{Aut_Ntaxa_Bias}{bias}
}}
\usage{
demo_observed_values
}
\description{
A dataset containing demo observed values
}
\keyword{datasets}
