% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_extensions.R
\name{gt_add_column}
\alias{gt_add_column}
\title{Add a column to a \code{gt}}
\usage{
gt_add_column(tab, .col, .val, ...)
}
\arguments{
\item{tab}{A \code{gt} table with class \code{gt_tbl}}

\item{.col}{column name}

\item{.val}{value to add}

\item{...}{additional arguments to pass to \link[dplyr:reexports]{dplyr::add_row}}
}
\value{
A \link[gt:gt]{gt::gt}
}
\description{
Add a column to a \code{gt}
}
\examples{
rict(wv) |> gt_add_column(r, 3)
}
