% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_extensions.R
\name{gt_get_data}
\alias{gt_get_data}
\title{Extract data from a \code{gt}}
\usage{
gt_get_data(tab)
}
\arguments{
\item{tab}{A \code{gt} table with class \code{gt_tbl}}
}
\value{
A \link[dplyr:reexports]{dplyr::tibble}
}
\description{
Extract data from a \code{gt}
}
\examples{
rict(wv) |> gt_get_data()
}
