% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_extensions.R
\name{gt_hide_lists}
\alias{gt_hide_lists}
\title{Hide List Columns in \code{gt}}
\usage{
gt_hide_lists(tab)
}
\arguments{
\item{tab}{A \code{gt} table with class \code{gt_tbl}}
}
\value{
A \link[gt:gt]{gt::gt}
}
\description{
Hide List Columns in \code{gt}
}
\examples{
wv \%>\% gt::gt() \%>\% gt_hide_lists()
}
