% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_plot_compactness.R
\name{gt_plot_compactness}
\alias{gt_plot_compactness}
\title{Add Compactness Metric Plots to a \code{gt}}
\usage{
gt_plot_compactness(
  tab,
  shp,
  plan,
  measures = guess_comp(tab),
  height = 200,
  ...
)
}
\arguments{
\item{tab}{A \code{gt} table with class \code{gt_tbl}}

\item{shp}{An \code{sf} object}

\item{plan}{A numeric vector with one entry for each precinct in \code{shp}.}

\item{measures}{A character vector indicating which measures to plot. Uses \code{guess_comp()} if not supplied.}

\item{height}{height, in pixels, of each image. Default is 200.}

\item{...}{additional arguments. Not currently passed on.}
}
\value{
A \link[gt:gt]{gt::gt}
}
\description{
Add Compactness Metric Plots to a \code{gt}
}
\examples{
rict(wv_plans, 'cd_2020') \%>\%
  gt_plot_compactness(wv, wv$cd_2020)
}
