% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_plot_sf.R
\name{gt_plot_sf}
\alias{gt_plot_sf}
\title{Add \code{sf} Geometry Plots to a \code{gt}}
\usage{
gt_plot_sf(tab, name, height = 100, ...)
}
\arguments{
\item{tab}{A \code{gt} table with class \code{gt_tbl}}

\item{name}{Name for \code{sf} column in \code{gt}}

\item{height}{height, in pixels, of each image. Default is 100}

\item{...}{additional arguments passed on to \code{geom_sf()}}
}
\value{
A \link[gt:gt]{gt::gt}
}
\description{
Add \code{sf} Geometry Plots to a \code{gt}
}
\examples{
wv_dist <- wv \%>\% dplyr::group_by(cd_2020) \%>\% dplyr::summarize()
gt::gt(wv_dist) \%>\% gt_plot_sf()
}
