% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_color_party.R
\docType{data}
\name{partisan}
\alias{partisan}
\title{Partisan Color Palette}
\format{
A character vector of 15 hex colors with class \code{"palette"}.
}
\usage{
partisan
}
\description{
A red-to-blue 15-color palette for partisan vote share, where 0 is fully
Republican and 1 is fully Democratic.
}
\examples{
partisan
}
\keyword{datasets}
