% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compactness.R
\name{plot_compactness}
\alias{plot_compactness}
\title{Create Plots for Common Compactness Metrics}
\usage{
plot_compactness(
  shp,
  plan,
  measure = c("Polsby Popper", "Schwartzberg", "Reock", "Convex Hull", "Length Width",
    "Skew", "Box Reock"),
  fill_color = "deeppink"
)
}
\arguments{
\item{shp}{An \code{sf} object}

\item{plan}{A numeric vector with one entry for each precinct in \code{shp}.}

\item{measure}{A character indicating which measure to plot. Default is \code{'Polsby Popper'}.}

\item{fill_color}{hex or color name to fill the shape. A second entry can be provided
to fill the background.}
}
\value{
list of \code{ggplot2} plots
}
\description{
Create Plots for Common Compactness Metrics
}
\examples{
plot_compactness(wv, wv$cd_2020)
}
