% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict-package.R
\docType{package}
\name{rict-package}
\alias{rict-package}
\title{rict: Redistricting in Clean Tables}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a suite of tools to create tables that accompany maps. The focus is clean, informative tables for electoral outcomes, compactness, and other district-level quantities. Most tools are aimed at the redistricting context, but are broadly applicable to other electoral data.
}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{ctkenny@proton.me} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
