% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict-classifcation.R
\name{rict}
\alias{rict}
\title{Run the River Invertebrate Classification Tool (RICT)}
\usage{
rict(data = NULL, model = "physical", area = "gb", year_type = "multi")
}
\arguments{
\item{data}{Dataframe of observed environmental values
\describe{
  \item{SITE}{Site identifier}
  \item{Waterbody}{Water body identifier}
  ...
}}

\item{model}{Model to use either 'gis' or 'physical'}

\item{area}{Area of UK, either 'ni' (Northern Ireland) or 'gb' (Great Britain)}

\item{year_type}{"single" or "multi" depending if multi-year classification
required - default is "multi"}
}
\value{
Dataframe of classification results
}
\description{
The \code{rict} function classifies observed environmental values.
}
\examples{
\dontrun{
results <- rict(demo_observed_values)
}
}
