% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.R
\name{rict_boundary}
\alias{rict_boundary}
\title{Display boundary information in a table}
\usage{
rict_boundary(map, plan, seam, adj_col = "adj", as_gt = TRUE)
}
\arguments{
\item{map}{A \code{redist_map} or \code{sf} object.}

\item{plan}{Column in \code{map} or vector of district assignments.}

\item{seam}{Pair of districts in \code{plan} to focus on.}

\item{adj_col}{Name of column in \code{map} that contains adjacency information.}

\item{as_gt}{Logical. Should output be a \code{gt} table? Default: \code{TRUE}.}
}
\value{
a \code{gt_tbl} if \code{as_gt = TRUE}, otherwise a \link[tibble:tibble]{tibble::tibble}
}
\description{
Display boundary information in a table
}
\examples{
rict_boundary(map = wv, plan = wv$cd_2020, seam = c(1, 2))
}
