% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-classification.R
\name{rict_classify}
\alias{rict_classify}
\title{Calculate classification}
\usage{
rict_classify(data = NULL, year_type = "multi")
}
\arguments{
\item{data}{Dataframe of predicted endgroups values from
'rict__predict' function}

\item{year_type}{"single" or "multi" depending if multi-year classification
required - default is "multi"}
}
\value{
Dataframe of classification results
}
\description{
Calculate classification
}
\examples{
\dontrun{
predictions <- rict_predict(rict_observed = rict::demo_rict_observed)
classifications <- rict_classify(predictions)
}

}
