% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{rict_compactness}
\alias{rict_compactness}
\title{Display compactness measures in a table}
\usage{
rict_compactness(
  map,
  plan,
  measures = list(comp_polsby = redistmetrics::comp_polsby, comp_schwartz =
    redistmetrics::comp_schwartz, comp_reock = redistmetrics::comp_reock, comp_ch =
    redistmetrics::comp_ch),
  as_gt = TRUE
)
}
\arguments{
\item{map}{A \code{redist_map} or \code{sf} object.}

\item{plan}{Column in \code{map} or vector of district assignments.}

\item{measures}{a list of named functions to score compactness}

\item{as_gt}{Logical. Should output be a \code{gt} table? Default: \code{TRUE}.}
}
\value{
a \code{gt_tbl} if \code{as_gt = TRUE}, otherwise a \link[tibble:tibble]{tibble::tibble}
}
\description{
Display compactness measures in a table
}
\examples{
rict_compactness(map = wv, plan = wv$cd_2020)
}
