% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{rict_component}
\alias{rict_component}
\title{Display population data by administrative unit in a table}
\usage{
rict_component(map, plan, admin, as_gt = TRUE)
}
\arguments{
\item{map}{A \code{redist_map} or \code{sf} object.}

\item{plan}{Column in \code{map} or vector of district assignments.}

\item{admin}{column names in \code{map} without \code{NA} values to calculate administrative splits for}

\item{as_gt}{Logical. Should output be a \code{gt} table? Default: \code{TRUE}.}
}
\value{
a \code{gt_tbl} if \code{as_gt = TRUE}, otherwise a \link[tibble:tibble]{tibble::tibble}
}
\description{
Display population data by administrative unit in a table
}
\examples{
rict_component(map = wv, plan = wv$cd_2020, admin = 'county')
}
