% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{rict_contiguity}
\alias{rict_contiguity}
\title{Display contiguity info in a table}
\usage{
rict_contiguity(map, plan, adj_col = "adj", as_gt = TRUE)
}
\arguments{
\item{map}{A \code{redist_map} or \code{sf} object.}

\item{plan}{Column in \code{map} or vector of district assignments.}

\item{adj_col}{Name of column in \code{map} that contains adjacency information.}

\item{as_gt}{Logical. Should output be a \code{gt} table? Default: \code{TRUE}.}
}
\value{
a \code{gt_tbl} if \code{as_gt = TRUE}, otherwise a \link[tibble:tibble]{tibble::tibble}
}
\description{
Display contiguity info in a table
}
\examples{
rict_contiguity(map = wv, plan = wv$cd_2020)
}
