% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{rict_demographics}
\alias{rict_demographics}
\title{Display demographic data in a table}
\usage{
rict_demographics(map, plan, normalize = TRUE, as_gt = TRUE)
}
\arguments{
\item{map}{A \code{redist_map} or \code{sf} object.}

\item{plan}{Column in \code{map} or vector of district assignments.}

\item{normalize}{Logical. Should columns be normalized to percentages? Default: \code{TRUE}.}

\item{as_gt}{Logical. Should output be a \code{gt} table? Default: \code{TRUE}.}
}
\value{
a \code{gt_tbl} if \code{as_gt = TRUE}, otherwise a \link[tibble:tibble]{tibble::tibble}
}
\description{
Display demographic data in a table
}
\examples{
rict_demographics(map = wv, plan = wv$cd_2020)
}
