% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{rict_population}
\alias{rict_population}
\title{Display population parity in a table}
\usage{
rict_population(map, plan, as_gt = TRUE)
}
\arguments{
\item{map}{A \code{redist_map} or \code{sf} object.}

\item{plan}{Column in \code{map} or vector of district assignments.}

\item{as_gt}{Logical. Should output be a \code{gt} table? Default: \code{TRUE}.}
}
\value{
a \code{gt_tbl} if \code{as_gt = TRUE}, otherwise a \link[tibble:tibble]{tibble::tibble}
}
\description{
Display population parity in a table
}
\examples{
rict_population(map = wv, plan = wv$cd_2020)
}
