% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict-predict.R
\name{rict_predict}
\alias{rict_predict}
\title{Calculate River Invertebrate Classification Tool (RICT) predictions}
\usage{
rict_predict(data = NULL, all_indices = FALSE, taxa = FALSE)
}
\arguments{
\item{data}{Dataframe of predictive environmental values
\describe{
  \item{SITE}{Site identifier}
  \item{Waterbody}{Water body identifier}
  \item{SPR_SEASON_ID}{Spring ID}
  \item{SUM_SEASON_ID}{Summary ID}
  \item{AUT_SEASON_ID}{Aut ID}
  \item{...}{etc}
}}

\item{all_indices}{Return all indices in output (default only returns WHPT indices).}

\item{taxa}{Return taxa predictions (default  returns indices).}
}
\value{
Dataframe of predicted biotic scores and probability of observed values
falling into each statistical grouping of rivers.
}
\description{
Calculate River Invertebrate Classification Tool (RICT) predictions
}
\examples{
\dontrun{
predictions <- rict_predict(demo_observed_values)
}
}
