% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{rict_splits}
\alias{rict_splits}
\title{Display splits data in a table}
\usage{
rict_splits(
  map,
  plan,
  admin = NULL,
  subadmin = NULL,
  total = admin,
  multi = admin,
  as_gt = TRUE
)
}
\arguments{
\item{map}{A \code{redist_map} or \code{sf} object.}

\item{plan}{Column in \code{map} or vector of district assignments.}

\item{admin}{column names in \code{map} without \code{NA} values to calculate administrative splits for}

\item{subadmin}{column names in \code{map} with \code{NA} values to calculate administrative splits for}

\item{total}{column names in \code{map} without \code{NA} values to calculate total splits for}

\item{multi}{column names in \code{map} without \code{NA} values to calculate multi-splits for}

\item{as_gt}{Logical. Should output be a \code{gt} table? Default: \code{TRUE}.}
}
\value{
a \code{gt_tbl} if \code{as_gt = TRUE}, otherwise a \link[tibble:tibble]{tibble::tibble}
}
\description{
Display splits data in a table
}
\examples{
rict_splits(map = wv, plan = wv$cd_2020, admin = 'state')
}
