% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict-validate.R
\name{rict_validate}
\alias{rict_validate}
\title{Validate observed values}
\usage{
rict_validate(data = NULL, model = "physical", area = "gb")
}
\arguments{
\item{data}{dataframe of observed envornmental variables
\describe{
  \item{SITE}{Site identifier}
  \item{Waterbody}{Water body identifier}
  ...
}}

\item{model}{Specify model to apply correct validation rules, default is
"physical".}

\item{area}{Specify area of UK to apply localised validation rules, default
is "gb". Accepts "ni" (Northern Ireland) or "gb" (Great Britain).}
}
\value{
List of three dataframes: 1. warnings_failings, 2. this_failing 3.
  valid 'data'.
}
\description{
Low level function to validate input `data`. Returns list of dataframes
containing fails, warnings and passing observed values. The validation rules
adjust based on the `model` and `area` arguments. Here is a summary of
checks:
\enumerate{
 \item Input 'data' exists
 \item Input is dataframe
 \item Required columns are present
 \item Columns have correct class
 \item Conditional columns if present are correct class
 \item Where multiple classes are allowed, convert columns to
 standardised class for example integer to numeric.
 \item Suspected input errors are corrected for example `0` slope.
 \item Additional columns/variables calculated for example mean temperate.
 \item Logs failures and warnings applied using `validation_rules` table.
 \item
 \item Replacements made to very low values for example
 \item Returns dataframe with passing values, warnings, fails
}
See `validation` vignette for full details.
}
\examples{
\dontrun{
validations <- rict_validate(demo_observed_values)
}
}
