% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{tally_pop}
\alias{tally_pop}
\title{Tally population columns by district}
\usage{
tally_pop(
  map,
  plan,
  pop_cols = dplyr::starts_with("pop_"),
  pop = "pop",
  normalize = FALSE
)
}
\arguments{
\item{map}{A \code{redist_map} or \code{sf} object.}

\item{plan}{Column in \code{map} or vector of district assignments.}

\item{pop_cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> columns to tally.
Default: columns starting with \code{"pop_"}.}

\item{pop}{Name of the total population column. Default: \code{"pop"}.}

\item{normalize}{Logical. Normalize subcategory columns by total? Default: \code{FALSE}.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with one row per district.
}
\description{
Tally population columns by district
}
\examples{
tally_pop(wv, wv$cd_2020, normalize = TRUE)
}
