% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{tally_vap}
\alias{tally_vap}
\title{Tally voting-age population columns by district}
\usage{
tally_vap(
  map,
  plan,
  vap_cols = dplyr::starts_with("vap_"),
  vap = "vap",
  normalize = FALSE
)
}
\arguments{
\item{map}{A \code{redist_map} or \code{sf} object.}

\item{plan}{Column in \code{map} or vector of district assignments.}

\item{vap_cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> columns to tally.
Default: columns starting with \code{"vap_"}.}

\item{vap}{Name of the total VAP column. Default: \code{"vap"}.}

\item{normalize}{Logical. Normalize subcategory columns by total? Default: \code{FALSE}.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with one row per district.
}
\description{
Tally voting-age population columns by district
}
\examples{
tally_vap(wv, wv$cd_2020, normalize = TRUE)
}
