% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{wv}
\alias{wv}
\title{West Virginia Geographic Data}
\format{
\code{redist_map} object
\describe{
\item{\code{GEOID}}{US Census Geographic Identifier}
\item{\code{NAME}}{County name}
\item{\code{pop_hisp}}{Hispanic population}
\item{\code{pop_white}}{White, non-Hispanic population}
\item{\code{pop_black}}{Black, non-Hispanic population}
\item{\code{pop_aian}}{American Indian and Alaskan Native, non-Hispanic population}
\item{\code{pop_asian}}{Asian, non-Hispanic population}
\item{\code{pop_nhpi}}{Native Hawaiian and Pacific Islander, non-Hispanic population}
\item{\code{pop_other}}{Other, non-Hispanic population}
\item{\code{pop_two}}{Two or More Races, non-Hispanic population}
\item{\code{vap}}{voting age population}
\item{\code{vap_hisp}}{Hispanic voting age population}
\item{\code{vap_white}}{White, non-Hispanic voting age population}
\item{\code{vap_black}}{Black, non-Hispanic voting age population}
\item{\code{vap_aian}}{American Indian and Alaskan Native, non-Hispanic voting age population}
\item{\code{vap_asian}}{Asian, non-Hispanic voting age population}
\item{\code{vap_nhpi}}{Native Hawaiian and Pacific Islander, non-Hispanic voting age population}
\item{\code{vap_other}}{Other, non-Hispanic voting age population}
\item{\code{vap_two}}{Two or More Races, non-Hispanic voting age population}
\item{\code{cd_2010}}{2010 congressional district lines smoothed to the county level}
\item{\code{cd_2020}}{2020 congressional district lines}
\item{\code{pre_20_dem_bid}}{votes for Biden 2020, President (D)}
\item{\code{pre_20_rep_tru}}{votes for Trump 2020, President (R)}
\item{\code{arv_20}}{average Republican vote in 2020}
\item{\code{adv_20}}{average Democratic vote in 2016}
\item{\code{nrv}}{normal Republican vote}
\item{\code{ndv}}{normal Democratic vote}
\item{\code{adj}}{adjacency list, zero-indexed}
\item{\code{sample_1}}{random sampled plan from \code{redist} 50 states project}
\item{\code{sample_2}}{random sampled plan from \code{redist} 50 states project}
\item{\code{sample_3}}{random sampled plan from \code{redist} 50 states project}
\item{\code{sample_4}}{random sampled plan from \code{redist} 50 states project}
\item{\code{sample_5}}{random sampled plan from \code{redist} 50 states project}
\item{\code{sample_6}}{random sampled plan from \code{redist} 50 states project}
\item{\code{sample_7}}{random sampled plan from \code{redist} 50 states project}
\item{\code{sample_8}}{random sampled plan from \code{redist} 50 states project}
\item{\code{sample_9}}{random sampled plan from \code{redist} 50 states project}
\item{\code{sample_10}}{random sampled plan from \code{redist} 50 states project}
\item{\code{geometry}}{sf geometry}
}
}
\description{
This file contains demographic, partisan, and geographic data for West Virginia at the
county level.
}
\examples{
data(wv)
}
\concept{data}
