% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{wv_plans}
\alias{wv_plans}
\title{West Virginia Redistricting Plans}
\format{
\code{redist_plans} object
\describe{
\item{\code{draw}}{draw identifier}
\item{\code{district}}{district number}
\item{\code{total_pop}}{Total population}
\item{\code{total_vap}}{voting age population}
\item{\code{plan_dev}}{Maximum deviation from perfect population parity}
\item{\code{comp_edge}}{Fraction of Edges Kept compactness}
\item{\code{comp_polsby}}{Polsby Popper compactness}
\item{\code{pop_white}}{White, non-Hispanic population}
\item{\code{pop_black}}{Black, non-Hispanic population}
\item{\code{pop_hisp}}{Hispanic population}
\item{\code{pop_aian}}{American Indian and Alaskan Native, non-Hispanic population}
\item{\code{pop_asian}}{Asian, non-Hispanic population}
\item{\code{pop_nhpi}}{Native Hawaiian and Pacific Islander, non-Hispanic population}
\item{\code{pop_other}}{Other, non-Hispanic population}
\item{\code{pop_two}}{Two or More Races, non-Hispanic population}
\item{\code{vap_hisp}}{Hispanic voting age population}
\item{\code{vap_white}}{White, non-Hispanic voting age population}
\item{\code{vap_black}}{Black, non-Hispanic voting age population}
\item{\code{vap_aian}}{American Indian and Alaskan Native, non-Hispanic voting age population}
\item{\code{vap_asian}}{Asian, non-Hispanic voting age population}
\item{\code{vap_nhpi}}{Native Hawaiian and Pacific Islander, non-Hispanic voting age population}
\item{\code{vap_other}}{Other, non-Hispanic voting age population}
\item{\code{vap_two}}{Two or More Races, non-Hispanic voting age population}
\item{\code{pre_20_dem_bid}}{votes for Biden 2020, President (D)}
\item{\code{pre_20_rep_tru}}{votes for Trump 2020, President (R)}
\item{\code{arv_20}}{average Republican vote in 2020}
\item{\code{adv_20}}{average Democratic vote in 2020}#' \item{\code{nrv}}{normal Republican vote}
\item{\code{ndv}}{normal Democratic vote}
\item{\code{ndshare}}{normal Democratic share in the district}
\item{\code{e_dvs}}{expected Democratic share in the district}
\item{\code{pr_dem}}{proportion of districts where Democrats win reconstructed elections}
\item{\code{e_dem}}{expected number of Democratic seats}
\item{\code{pbias}}{partisan bias}
\item{\code{egap}}{efficiency gap}
}
}
\description{
This file contains 10 sampled plans from the ALARM Project 50 states project
and the 2020 congressional plan for WV.
}
\examples{
data(wv_plans)
}
\concept{data}
