% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-classification.R
\name{calcClassification}
\alias{calcClassification}
\title{This experiment is a multi-year classification Monte-Carlo simulation of predictions}
\usage{
calcClassification(predictions)
}
\arguments{
\item{predictions}{Dataframe of predicted endgroups values from calcPredictions function}
}
\value{
Dataframe of MultiYear classifications
}
\description{
This experiment is a multi-year classification Monte-Carlo simulation of predictions
}
\examples{
\dontrun{
predictions <- calcPredictions(observed_values = rict::demo_observed_values)
classifications <- calcClassification(predictions)
}
}
