% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRict.R
\name{calcRict}
\alias{calcRict}
\title{Calculate River Invertebrate Classification Tool(RICT) predictions}
\usage{
calcRict(environmentalVariables, endPointMeans)
}
\arguments{
\item{environmentalVariables}{Dataframe of observed environmental values}

\item{endPointMeans}{Dataframe of endPointMeans}
}
\value{
Dataframe of predicted values
}
\description{
Calculate River Invertebrate Classification Tool(RICT) predictions
}
\examples{
\dontrun{
environmentalVariables <- read.csv(file="inst/extdat/fba-test-WHPT.csv")
endPointMeans <- read.csv(file="inst/extdat/End_Group_Means.csv")
output <- calcRict(environmentalVariables=environmentalVariables, endPointMeans=endPointMeans)
}
}
