% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare the statistical differences between classifications results}
\usage{
compare(data = NULL, a = NULL, b = NULL)
}
\arguments{
\item{data}{Data frame containing at least three columns.  A 'RESULT' and 'SITE' column
plus at least one column with simulated eqrs. For
example see `demo_eqr_values` - this is nested but flat dataframe will work too.}
}
\value{
Dataframe of compare results
}
\description{
Low-level function to assess whether there is a real
difference in EQR values and/or status class between simulated EQR results.
What this function does:
\enumerate{
 \item Run probability of difference
 \item Enumerate probability of difference
}
}
\examples{
\dontrun{
compare_results <- compare(demo_eqr_values)
}

}
