% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict.R
\name{rict}
\alias{rict}
\title{Run the River Invertebrate Classification Tool (RICT)}
\usage{
rict(data = NULL, year_type = "multi", store_eqrs = F)
}
\arguments{
\item{data}{Dataframe of observed environmental values
\describe{
 \item{SITE}{Site identifier}
 \item{Waterbody}{Water body identifier}
  }}

\item{year_type}{"single" or "multi" depending if multi-year classification
required - default is "multi".}

\item{store_eqrs}{TRUE/FALSE if set to TRUE function will return simulate
EQRs only}
}
\value{
Dataframe of classification results (or simulated EQRs if store_eqrs
  = TRUE)
}
\description{
The \code{rict} function classifies observed environmental values.
}
\examples{
\dontrun{
results <- rict(demo_observed_values)
}
}
