% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict.R
\name{rict}
\alias{rict}
\title{Run the River Invertebrate Classification Tool (RICT)}
\usage{
rict(data = NULL, year_type = "multi", store_eqrs = FALSE)
}
\arguments{
\item{data}{Dataframe of observed environmental values for example of input
 data see \code{\link{demo_observed_values}} or
 \code{\link{demo_gis_values_log}}.
\describe{
 \item{SITE}{Site identifier}
 \item{Waterbody}{Water body identifier}
 \item{...}{...}
  }}

\item{year_type}{"single" or "multi" depending if multi-year classification
required - default is "multi".}

\item{store_eqrs}{TRUE/FALSE if set to TRUE function will return simulate
EQRs only}
}
\value{
Dataframe of classification results (and simulated EQRs if store_eqrs
  = TRUE)
}
\description{
The \code{rict} function classifies observed environmental values.
}
\examples{
\dontrun{
results <- rict(demo_observed_values)
}
}
\seealso{
\code{\link{rict_predict}} to run prediction (not classification)
}
