% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict.R
\name{rict}
\alias{rict}
\title{Title}
\usage{
rict(x, plan, ...)
}
\arguments{
\item{x}{redist_map or redist_plans}

\item{plan}{integer vector or draw name}

\item{...}{other arguments}
}
\value{
A \link[gt:gt]{gt::gt}
}
\description{
Title
}
\examples{
data(wv_plans)
rict(wv)
}
