% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict-app.R
\name{rict_app}
\alias{rict_app}
\title{Open RICT web app}
\usage{
rict_app()
}
\description{
Open RICT as an interactive shiny web app.
}
\details{
\code{rict_app()} opens RICT as an interactive shiny app. RICT will
automatically detect if the input file contains NI or GB grid references and
apply the correct model including if GIS / geological predictors are used
(currently only available for GB).

Using the app:
\enumerate{
 \item Select your required year type (multi/single) for classification and
 type of prediction for instance include taxa or all indices.
 \item Click the Browse for .csv file button to select and upload a data file.
 \item View and download results.
 }
}
\examples{
\dontrun{
rict_app()
}

}
