% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict-classify.R
\name{rict_classify}
\alias{rict_classify}
\title{Calculate classification}
\usage{
rict_classify(
  data = NULL,
  year_type = "multi",
  store_eqrs = FALSE,
  n_runs = 10000
)
}
\arguments{
\item{data}{Dataframe of predicted endgroups values from
\code{\link{rict__predict}} function}

\item{year_type}{"single" or "multi" depending if multi-year classification
required - default is "multi"}

\item{store_eqrs}{Boolean to signal if simulate EQRs should be stored. If
TRUE, EQRs are stored allowing `rict_compare` to compare EQR results}

\item{n_runs}{Number of simulations, default 10000.}
}
\value{
Dataframe of classification results
}
\description{
Calculate classification
}
\examples{
\dontrun{
predictions <- rict_predict(demo_observed_values)
classifications <- rict_classify(predictions)
}

}
