% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict-compare.R
\name{rict_compare}
\alias{rict_compare}
\title{Compare the Statistical Differences between Classifications Results}
\usage{
rict_compare(
  results_a = NULL,
  results_b = NULL,
  eqr_metrics = c("AVG_NTAXA", "AVG_ASPT")
)
}
\arguments{
\item{results_a}{Data frame output from `rict(store_eqrs = T)`.}

\item{results_b}{Data frame output from `rict(store_eqrs = T)`}

\item{list}{of eqr_metrics to compare, default is average year ASPT and NTAXA
metric}
}
\value{
Dataframe of compare results
}
\description{
The Compare function allows the user to assess whether there is a real
difference in EQR values and/or status class between a pair of results and/or
sites and/or time periods.
}
\details{
By default detects if upstream / downstream comparison required. This is
indicate by 'a' sites not included in 'b' sites or else it detects
if sites are paired between both datasets. For instance, if comparing season
to season or year to year pairs.
}
\examples{
\dontrun{
results_a <- rict(demo_observed_values[1, ], store_eqrs = T, year_type = "single")
results_b <- rict(demo_gis_values[1, ], store_eqrs = T, year_type = "single", model = "gis")
compare <- rict_compare(results_a, results_b)
}

}
