% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict-compare.R
\name{rict_compare}
\alias{rict_compare}
\title{Compare the statistical differences between classifications results}
\usage{
rict_compare(data = NULL, eqrs = c("ASPT", "NTAXA"))
}
\arguments{
\item{data}{Standard input data - TBC: default compare site with same name or
extra column indicating pairs??}

\item{eqrs}{list of eqr metrics to compare for example ASPT, NTAXA, MINTA}
}
\value{
Dataframe of compare results
}
\description{
The Compare function allows the user to assess whether there is a real
difference in EQR values and/or status class between a pair of results and/or
sites and/or time periods. What this function does:
\enumerate{
 \item Runs `rict_predict`.
 \item Runs  `rict_classify`
 \item Stores EQI (correct for bias) as well as normal classification outputs.
 \item When all results have run - check which sites need to be compareds.
}
}
\examples{
\dontrun{
compare <- rict_compare(demo_observed_values)
}
}
