% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict-compare.R
\name{rict_compare}
\alias{rict_compare}
\title{Compare the statistical significance of differences between classifications}
\usage{
rict_compare(data = rict::demo_observed_values)
}
\arguments{
\item{data}{Standard input data - TBC: default compare site with same name or extra column indicating pairs??
Run prediction
Run classification only until 10000 EQR have been generated?
Store EQI (correct for bias) as well as normal classification outputs? This could be a lot if big run?
When all samples have run - check which sites need to be compared}
}
\description{
The Compare function allows the user to assess whether there is a real
difference in EQR values and/or status class between a pair of samples and/or
sites and/or time periods
}
\details{
What this does:
}
