% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-predictions.R
\name{rict_predict}
\alias{rict_predict}
\title{Calculate River Invertebrate Classification Tool (RICT) predictions}
\usage{
rict_predict(data = NULL, model = "physical", area = "gb")
}
\arguments{
\item{data}{Dataframe of predictive environmental values
\describe{
  \item{SITE}{Site identifier}
  \item{Waterbody}{Water body identifier}
  \item{SPR_SEASON_ID}{Spring ID}
  \item{SUM_SEASON_ID}{Summary ID}
  \item{AUT_SEASON_ID}{Aut ID}
  \item{...}{etc}
}}

\item{model}{Validate observed values based on rules required for model selected.}

\item{area}{Area of UK, either 'ni' (Northern Ireland) or 'gb' (Great Britain)}
}
\value{
Dataframe of predicted biotic scores and probability of observed values
falling into each statistical grouping of rivers.
}
\description{
Calculate River Invertebrate Classification Tool (RICT) predictions
}
\examples{
\dontrun{
predictions <- rict_predict(data = rict::demo_observed_values)
}
}
