% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict-validate.R
\name{rict_validate}
\alias{rict_validate}
\title{Validate observed values}
\usage{
rict_validate(data = NULL, temp_override = F)
}
\arguments{
\item{data}{dataframe of observed environmental variables
\describe{
  \item{SITE}{Site identifier}
  \item{Waterbody}{Water body identifier}
  ...
}}

\item{temp_override}{If set to TRUE - user provided temperatures in columns
TMEAN and TRANGE will ovedrride calculated temperature generated from
NGR.This allows research on responses to temperature.}
}
\value{
List of three dataframes: 1. warnings_failings, 2. this_failing 3.
  valid 'data'.
}
\description{
Low level function to validate input `data`. Returns list of dataframes
containing fails/warnings and passing observed values. The function detects
which model (gis or physical) and area (GB & NI) and applies validation rules
as required. Here is a summary of checks:
\enumerate{
 \item Input 'data' exists
 \item Input is dataframe
 \item Required columns are present
 \item Columns have correct class
 \item Conditional columns if present are correct class
 \item Where multiple classes are allowed, convert columns to
 standardised class for example integer to numeric.
 \item Additional columns/variables calculated for example mean temperate.
 \item Logs failures and warnings applied using `validation_rules` table.
 \item
 \item Replacements made to very low values for example
 \item Returns dataframe with passing values, warnings, fails
}
See `validation` vignette for full details.
}
\examples{
\dontrun{
validations <- rict_validate(demo_observed_values)
}
}
