% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict-validate.R
\name{rict_validate}
\alias{rict_validate}
\title{Validate observed values}
\usage{
rict_validate(data = NULL, row = FALSE)
}
\arguments{
\item{data}{dataframe of observed environmental variables
\describe{
  \item{SITE}{Site identifier}
  \item{Waterbody}{Water body identifier}
  ...
}}

\item{row}{Boolean - if set to `TRUE` returns the row number from the input file
(data) for each check. This makes linking checks (fails/warns etc) to the
associated row in the input data easier. This is more relevant if multiple
samples from the same site and year are input as separate row to the
`rict_validate`. In this case, SITE and YEAR are not enough to link
validation checks to specific rows in the input data.}
}
\description{
Low level function to validate input `data`. Returns list of dataframes
containing fails/warnings/replacements and passing observed values. The
function detects which model (gis or physical) and area (GB & NI) and applies
validation rules as required. Here is a summary of checks:
\enumerate{
 \item Input 'data' exists.
 \item Input is dataframe.
 \item Required columns are present.
 \item Columns have correct class.
 \item Conditional columns if present are correct class.
 \item Where multiple classes are allowed, convert columns to
 standardised class for example integer to numeric.
 \item Additional columns/variables calculated for example mean temperate.
 \item Logs failures and warnings applied using `validation_rules` table.
 \item Replace values if value is less than the ‘overall’ minimum value.
 \item Returns dataframe with passing values, warnings, fails.
}
}
\examples{
\dontrun{
validations <- rict_validate(demo_observed_values,  row == TRUE)
}
}
